Brander/Back/requirement.txt
openai
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
